Remove unused hooks directory
