+++ package.json
@@ -89,10 +89,6 @@
     "type-fest": "^3.10.0",
     "typescript": "^4.7.4",
     "webpack": "^5.79.0",
-    "webpack-merge": "^5.8.0",
-    "react-syntax-highlighter": "^15.5.0"
-  },
-  "engines": {
-    "node": ">=16"
+    "webpack-merge": "^5.8.0"
   }
 }
