+++ packages/forklift-console-plugin/locales/en/plugin__forklift-console-plugin.json
@@ -26,7 +26,6 @@
   "Actions": "Actions",
   "Add mapping": "Add mapping",
   "Add source and target providers for the migration.": "Add source and target providers for the migration.",
-  "Adding hooks to the plan is optional. Hooks are contained in Ansible playbooks that can be run before or after the migration.": "Adding hooks to the plan is optional. Hooks are contained in Ansible playbooks that can be run before or after the migration.",
   "All discovered networks have been mapped to the default network.": "All discovered networks have been mapped to the default network.",
   "All discovered storages have been mapped to the default storage.": "All discovered storages have been mapped to the default storage.",
   "All networks detected on the selected VMs require a mapping.": "All networks detected on the selected VMs require a mapping.",
@@ -57,7 +56,6 @@
   "Category": "Category",
   "Certificate change detected": "Certificate change detected",
   "Clear all filters": "Clear all filters",
-  "Click the relevant buttons within the table for managing the plan hooks.": "Click the relevant buttons within the table for managing the plan hooks.",
   "Click the update credentials button to save your changes, button is disabled until a change is detected.": "Click the update credentials button to save your changes, button is disabled until a change is detected.",
   "Click the update mappings button to save your changes, button is disabled until a change is detected.": "Click the update mappings button to save your changes, button is disabled until a change is detected.",
   "Cluster": "Cluster",
@@ -119,7 +117,6 @@
   "Edit Controller Memory limit": "Edit Controller Memory limit",
   "Edit credentials": "Edit credentials",
   "Edit Default Transfer Network": "Edit Default Transfer Network",
-  "Edit hooks": "Edit hooks",
   "Edit mappings": "Edit mappings",
   "Edit Maximum concurrent VM migrations": "Edit Maximum concurrent VM migrations",
   "Edit migration plan target namespace": "Edit migration plan target namespace",
@@ -174,7 +171,6 @@
   "Hide from view": "Hide from view",
   "Hide values": "Hide values",
   "Hooks": "Hooks",
-  "Hooks data is loading, please wait.": "Hooks data is loading, please wait.",
   "Hooks for virtualization": "Hooks for virtualization",
   "Host": "Host",
   "Host cluster": "Host cluster",
@@ -249,7 +245,6 @@
   "New name was generated for the Storage Map due to naming conflict.": "New name was generated for the Storage Map due to naming conflict.",
   "NICs with empty NIC profile": "NICs with empty NIC profile",
   "No credentials found.": "No credentials found.",
-  "No hooks have been added to this migration plan.": "No hooks have been added to this migration plan.",
   "No inventory data available.": "No inventory data available.",
   "No Mapping found.": "No Mapping found.",
   "No NetworkMaps found.": "No NetworkMaps found.",
@@ -320,12 +315,10 @@
   "Pods": "Pods",
   "Pods not found": "Pods not found",
   "Post migration hook": "Post migration hook",
-  "Post-migration": "Post-migration",
   "Power state": "Power state",
   "Powered off": "Powered off",
   "Powered on": "Powered on",
   "Pre migration hook": "Pre migration hook",
-  "Pre-migration": "Pre-migration",
   "Precopy interval (minutes)": "Precopy interval (minutes)",
   "Preserve CPU model": "Preserve CPU model",
   "Preserve the CPU model and flags the VM runs with in its oVirt cluster.": "Preserve the CPU model and flags the VM runs with in its oVirt cluster.",
@@ -381,7 +374,6 @@
   "Skip certificate validation": "Skip certificate validation",
   "Snapshot polling interval (seconds)": "Snapshot polling interval (seconds)",
   "Something is wrong, the data was not loaded due to an error, please try to reload the page.": "Something is wrong, the data was not loaded due to an error, please try to reload the page.",
-  "Something is wrong, the hooks data was not loaded due to an error, please try to reload the page.": "Something is wrong, the hooks data was not loaded due to an error, please try to reload the page.",
   "Something is wrong, the secret was not loaded, please try to reload the page.": "Something is wrong, the secret was not loaded, please try to reload the page.",
   "source": "source",
   "Source Only": "Source Only",
@@ -415,7 +407,6 @@
   "The certificate is not a valid PEM-encoded X.509 certificate": "The certificate is not a valid PEM-encoded X.509 certificate",
   "The chosen provider is no longer available.": "The chosen provider is no longer available.",
   "The current certificate does not match the certificate fetched from URL. Manually validate the fingerprint before proceeding.": "The current certificate does not match the certificate fetched from URL. Manually validate the fingerprint before proceeding.",
-  "The edit hooks button is disabled if the plan started running and at least one virtual machine was migrated successfully.": "The edit hooks button is disabled if the plan started running and at least one virtual machine was migrated successfully.",
   "The edit mappings button is disabled if the plan started running and at least one virtual machine was migrated successfully.": "The edit mappings button is disabled if the plan started running and at least one virtual machine was migrated successfully.",
   "The interval in minutes for precopy. Default value is 60.": "The interval in minutes for precopy. Default value is 60.",
   "The interval in seconds for snapshot pooling. Default value is 10.": "The interval in seconds for snapshot pooling. Default value is 10.",
