--- packages/forklift-console-plugin/src/modules/NetworkMaps/utils/helpers/deepCopy.ts
@@ -1,3 +1,0 @@
-export function deepCopy<T>(obj: T): T {
-  return JSON.parse(JSON.stringify(obj));
-}
