+++ packages/forklift-console-plugin/src/modules/NetworkMaps/utils/helpers/index.ts
@@ -1,4 +1,3 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
-export * from './deepCopy';
 export * from './getNetworkMapPhase';
 // @endindex
