+++ packages/forklift-console-plugin/src/modules/NetworkMaps/views/details/components/MapsSection/state/reducer.ts
@@ -1,4 +1,4 @@
-import { deepCopy } from 'src/modules/NetworkMaps/utils';
+import { deepCopy } from 'src/utils';
 
 import { V1beta1NetworkMap, V1beta1NetworkMapSpecMap } from '@kubev2v/types';
