+++ packages/forklift-console-plugin/src/modules/Plans/utils/helpers/index.ts
@@ -1,5 +1,4 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
-export * from './deepCopy';
 export * from './getMigrationPhase';
 export * from './getMigrationVmsCounts';
 export * from './getPhaseLabel';
