+++ packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/Hooks/PlanHooks.tsx
@@ -1,87 +1,195 @@
-import React from 'react';
+import React, { useEffect, useReducer } from 'react';
+import { Base64 } from 'js-base64';
 import SectionHeading from 'src/components/headers/SectionHeading';
 import { useForkliftTranslation } from 'src/utils/i18n';
 
+import { CodeEditor } from '@openshift-console/dynamic-plugin-sdk';
 import {
-  HookModelGroupVersionKind,
-  PlanModelGroupVersionKind,
-  V1beta1Hook,
-  V1beta1Plan,
-} from '@kubev2v/types';
-import { useK8sWatchResource } from '@openshift-console/dynamic-plugin-sdk';
-import { PageSection } from '@patternfly/react-core';
+  Button,
+  Divider,
+  Flex,
+  FlexItem,
+  Form,
+  FormGroup,
+  HelperText,
+  HelperTextItem,
+  PageSection,
+  Switch,
+  TextInput,
+} from '@patternfly/react-core';
 
-import { PlanHooksSection } from './PlanHooksSection';
+import { Suspend } from '../../components';
 
-export type PlanHooksInitSectionProps = {
-  plan: V1beta1Plan;
-  loaded: boolean;
-  loadError: unknown;
-};
+import { usePlanHooks } from './hooks';
+import { formReducer, initialState } from './state';
+import { onUpdatePlanHooks } from './utils';
 
 export const PlanHooks: React.FC<{ name: string; namespace: string }> = ({ name, namespace }) => {
   const { t } = useForkliftTranslation();
 
-  const [plan, loaded, loadError] = useK8sWatchResource<V1beta1Plan>({
-    groupVersionKind: PlanModelGroupVersionKind,
-    namespaced: true,
+  const [plan, preHookResource, postHookResource, loaded, loadError] = usePlanHooks(
     name,
     namespace,
-  });
+  );
+
+  const [state, dispatch] = useReducer(
+    formReducer,
+    initialState(plan, preHookResource, postHookResource),
+  );
+
+  // Init state on outside changes
+  useEffect(() => {
+    dispatch({
+      type: 'INIT',
+      payload: initialState(plan, preHookResource, postHookResource),
+    });
+  }, [plan, preHookResource, postHookResource]);
+
+  // Handle user clicking "save"
+  async function onUpdate() {
+    onUpdatePlanHooks({ plan, preHookResource, postHookResource, dispatch, state });
+  }
+
+  const HooksTabAction = (
+    <Flex className="forklift-page-plan-details-vm-status__actions">
+      <FlexItem>
+        <Button
+          variant="primary"
+          onClick={onUpdate}
+          isDisabled={!state.hasChanges}
+          isLoading={state.isLoading}
+        >
+          {t('Update hooks')}
+        </Button>
+      </FlexItem>
+
+      <FlexItem>
+        <Button
+          variant="secondary"
+          isDisabled={!state.hasChanges}
+          onClick={() =>
+            dispatch({
+              type: 'INIT',
+              payload: initialState(plan, preHookResource, postHookResource),
+            })
+          }
+        >
+          {t('Cancel')}
+        </Button>
+      </FlexItem>
+    </Flex>
+  );
 
   return (
-    <>
-      <div>
-        <PageSection variant="light">
-          <SectionHeading text={t('Hooks')} />
-          <PlanHooksInitSection plan={plan} loaded={loaded} loadError={loadError} />
-        </PageSection>
-      </div>
-    </>
-  );
-};
+    <Suspend obj={plan} loaded={loaded} loadError={loadError}>
+      {state.alertMessage && <PageSection variant="light">{state.alertMessage}</PageSection>}
 
-const PlanHooksInitSection: React.FC<PlanHooksInitSectionProps> = (props) => {
-  const { t } = useForkliftTranslation();
-  const { plan } = props;
+      <PageSection
+        variant="light"
+        className="forklift-page-plan-details-vm-status__section-actions"
+      >
+        <SectionHeading text={t('Migration hook')} />
+        {HooksTabAction}
+        <Divider />
+      </PageSection>
 
-  // Retrieve all k8s Hooks
-  const [hooks, hooksLoaded, hooksLoadError] = useK8sWatchResource<V1beta1Hook[]>({
-    groupVersionKind: HookModelGroupVersionKind,
-    namespaced: true,
-    isList: true,
-    namespace: plan.metadata?.namespace,
-  });
+      <PageSection variant="light">
+        <SectionHeading text={t('Pre migration hook')} />
+        <Form>
+          <FormGroup label="Enable hook" isRequired fieldId="pre-hook-set">
+            <Switch
+              id="pre-hook-set"
+              label="Enable pre migration hook"
+              labelOff="Do not enable a pre migration hook"
+              isChecked={state.preHookSet}
+              onChange={(value) => dispatch({ type: 'PRE_HOOK_SET', payload: value })}
+            />
+          </FormGroup>
 
-  if (!hooksLoaded)
-    return (
-      <div>
-        <span className="text-muted">{t('Hooks data is loading, please wait.')}</span>
-      </div>
-    );
-
-  if (hooksLoadError)
-    return (
-      <div>
-        <span className="text-muted">
-          {t(
-            'Something is wrong, the hooks data was not loaded due to an error, please try to reload the page.',
+          {state.preHookSet && (
+            <>
+              <FormGroup label="Hook runner image" isRequired fieldId="pre-hook-image">
+                <TextInput
+                  value={state.preHook?.spec?.image}
+                  type="url"
+                  onChange={(value) => dispatch({ type: 'PRE_HOOK_IMAGE', payload: value })}
+                  aria-label="pre hook image"
+                />
+                <HelperText>
+                  <HelperTextItem>
+                    You can use a custom hook-runner image or specify a custom image, for example
+                    quay.io/konveyor/hook-runner .
+                  </HelperTextItem>
+                </HelperText>
+              </FormGroup>
+              <FormGroup label="Ansible playbook" fieldId="pre-hook-image">
+                <CodeEditor
+                  language="yaml"
+                  value={Base64.decode(state.preHook?.spec?.playbook || '')}
+                  onChange={(value) => dispatch({ type: 'PRE_HOOK_PLAYBOOK', payload: value })}
+                  minHeight="400px"
+                  showMiniMap={false}
+                />
+                <HelperText>
+                  <HelperTextItem>
+                    Optional: Ansible playbook. If you specify a playbook, the image must be
+                    hook-runner.
+                  </HelperTextItem>
+                </HelperText>
+              </FormGroup>
+            </>
           )}
-        </span>
-      </div>
-    );
+        </Form>
+      </PageSection>
 
-  // Search for the Plan k8s Hooks
-  const planHooks =
-    hooks?.filter((hook) =>
-      plan?.spec?.vms?.find((vm) =>
-        vm.hooks?.find(
-          (VMHook) =>
-            VMHook.hook.name === hook.metadata?.name &&
-            VMHook.hook.namespace === hook.metadata?.namespace,
-        ),
-      ),
-    ) || [];
+      <PageSection variant="light">
+        <SectionHeading text={t('Post migration hook')} />
+        <Form>
+          <FormGroup label="Enable hook" isRequired fieldId="post-hook-set">
+            <Switch
+              id="post-hook-set"
+              label="Enable post migration hook"
+              labelOff="Do not enable a post migration hook"
+              isChecked={state.postHookSet}
+              onChange={(value) => dispatch({ type: 'POST_HOOK_SET', payload: value })}
+            />
+          </FormGroup>
 
-  return <PlanHooksSection plan={plan} planHooks={planHooks} />;
+          {state.postHookSet && (
+            <>
+              <FormGroup label="Hook runner image" isRequired fieldId="post-hook-image">
+                <TextInput
+                  value={state.postHook?.spec?.image}
+                  type="url"
+                  onChange={(value) => dispatch({ type: 'POST_HOOK_IMAGE', payload: value })}
+                  aria-label="pre hook image"
+                />
+                <HelperText>
+                  <HelperTextItem>
+                    You can use a custom hook-runner image or specify a custom image, for example
+                    quay.io/konveyor/hook-runner .
+                  </HelperTextItem>
+                </HelperText>
+              </FormGroup>
+              <FormGroup label="Ansible playbook" fieldId="post-hook-image">
+                <CodeEditor
+                  language="yaml"
+                  value={Base64.decode(state.postHook?.spec?.playbook || '')}
+                  onChange={(value) => dispatch({ type: 'POST_HOOK_PLAYBOOK', payload: value })}
+                  minHeight="400px"
+                  showMiniMap={false}
+                />
+                <HelperText>
+                  <HelperTextItem>
+                    Optional: Ansible playbook. If you specify a playbook, the image must be
+                    hook-runner.
+                  </HelperTextItem>
+                </HelperText>
+              </FormGroup>
+            </>
+          )}
+        </Form>
+      </PageSection>
+    </Suspend>
+  );
 };
