+++ packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/Hooks/index.ts
@@ -1,4 +1,6 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './hooks';
 export * from './PlanHooks';
-export * from './PlanHooksSection';
+export * from './state';
+export * from './utils';
 // @endindex
