+++ packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/Hooks/state/index.ts
@@ -1,3 +1,4 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './initialState';
 export * from './reducer';
 // @endindex
