+++ packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/Hooks/utils/index.ts
