--- packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/SimpleHooks/index.ts
@@ -1,5 +1,0 @@
-// @index(['./*', /style/g], f => `export * from '${f.path}';`)
-export * from './SimpleHooks';
-export * from './state';
-export * from './utils';
-// @endindex
