+++ packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/index.ts
@@ -3,7 +3,6 @@ export * from './Details';
 export * from './Hooks';
 export * from './Mappings';
 export * from './Resources';
-export * from './SimpleHooks';
 export * from './VirtualMachines';
 export * from './YAML';
 // @endindex
