+++ packages/forklift-console-plugin/src/modules/StorageMaps/views/details/components/MapsSection/state/reducer.ts
@@ -1,4 +1,4 @@
-import { deepCopy } from 'src/modules/StorageMaps/utils';
+import { deepCopy } from 'src/utils';
 
 import { V1beta1StorageMap, V1beta1StorageMapSpecMap } from '@kubev2v/types';
