+++ packages/forklift-console-plugin/src/utils/deepCopy.ts
