+++ packages/forklift-console-plugin/src/utils/index.ts
@@ -1,0 +1,8 @@
+// @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './deepCopy';
+export * from './enums';
+export * from './fetch';
+export * from './i18n';
+export * from './resources';
+export * from './types';
+// @endindex
