+++ packages/forklift-console-plugin/src/utils/types.ts
@@ -1,34 +1,3 @@
-import { K8sGroupVersionKind } from '@openshift-console/dynamic-plugin-sdk';
-
-export type Condition = {
-  /** identification string for the condition */
-  type: string;
-  /** 'True' | 'False' indication for this condition */
-  status: string;
-  /** human readable message */
-  message: string;
-  /** higher order category this condition relates to */
-  category?: string;
-  /** Items related to this conditions */
-  items?: string[];
-  /** timestamp RFC 3339 */
-  lastTransitionTime?: string;
-  /** shorter computer friendly version of message */
-  reason?: string;
-};
-
-export type ResourceConsolePageProps = {
-  kind: string;
-  namespace: string;
-};
-
-export interface ProviderRef {
-  name: string;
-  gvk: K8sGroupVersionKind;
-  ready: boolean;
-  resolved: boolean;
-}
-
 export const ProviderStatusValues = [
   'ValidationFailed',
   'ConnectionFailed',
